[chore/#195] 패키지 분리
