+++ app/src/main/java/com/poti/android/domain/model/delivery/DeliveryInfo.kt
@@ -1,0 +1,9 @@
+package com.poti.android.domain.model.delivery
+
+data class DeliveryInfo(
+    val receiverName: String,
+    val zipcode: String,
+    val address: String,
+    val addressDetail: String,
+    val phoneNumber: String,
+)
