+++ app/src/main/java/com/poti/android/domain/model/party/PartyJoinInfo.kt
@@ -1,18 +1,12 @@
 package com.poti.android.domain.model.party
 
+import com.poti.android.domain.model.delivery.DeliveryInfo
+
 data class PartyJoinInfo(
     val partyId: Long,
     val shippingOptionId: Long,
     val deliveryInfo: DeliveryInfo,
     val joinItems: List<JoinOption>,
-)
-
-data class DeliveryInfo(
-    val receiverName: String,
-    val zipcode: String,
-    val address: String,
-    val addressDetail: String,
-    val phoneNumber: String,
 )
 
 data class JoinOption(
